requirements.txt
Flask==3.0.3
paho-mqtt==1.6.1
flask-socketio==5.3.6
